# docs/swagger.yaml
basePath: /v1
definitions:
  accounts.GetAccountResponse:
    properties:
      address:
        type: string
      birthday:
        type: string
      email:
        type: string
      firstName:
        type: string
      gender:
        $ref: '#/definitions/users.Gender'
      lastName:
        type: string
      photoUrl:
        type: string
    type: object
  accounts.UpdateAccountRequestPatch:
    properties:
      address:
        type: string
      birthday:
        type: string
      firstName:
        type: string
      gender:
        $ref: '#/definitions/users.Gender'
      lastName:
        type: string
    type: object
  accounts.UpdateAccountRequestPut:
    properties:
      address:
        type: string
      birthday:
        type: string
      firstName:
        type: string
      gender:
        allOf:
        - $ref: '#/definitions/users.Gender'
        enum:
        - male
        - female
      lastName:
        type: string
    required:
    - birthday
    - firstName
    - gender
    - lastName
    type: object
  accounts.UpdateAccountResponse:
    properties:
      address:
        type: string
      birthday:
        type: string
      email:
        type: string
      firstName:
        type: string
      gender:
        $ref: '#/definitions/users.Gender'
      lastName:
        type: string
      photoUrl:
        type: string
    type: object
...
  bands.BandResponse:
    description: Band response model
    properties:
      createdAt:
        type: string
      description:
        type: string
      genre:
        type: string
      id:
        type: integer
      name:
        type: string
      updatedAt:
        type: string
    type: object
  bands.CreateBandRequest:
    description: Create band request
    properties:
      description:
        type: string
      genre:
        maxLength: 100
        type: string
      name:
        maxLength: 255
        type: string
    required:
    - name
    type: object
  bands.ListBandsResponse:
    description: List bands response
    properties:
      items:
        items:
          $ref: '#/definitions/bands.BandResponse'
        type: array
    type: object
  bands.UpdateBandRequest:
    description: Update band request
    properties:
      description:
        type: string
      genre:
        maxLength: 100
        type: string
      name:
        maxLength: 255
        type: string
    type: object
...
  users.Gender:
    enum:
    - male
    - female
    type: string
    x-enum-varnames:
    - Male
    - Female
...
host: localhost:777
info:
  contact: {}
  description: This is a Concert booking application
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Concert booking API
  version: "1.0"
paths:
  /admin/v1/bands:
    get:
      description: Get a paginated list of bands
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Page size (default: 10, max: 100)'
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/bands.ListBandsResponse'
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
      summary: List bands
      tags:
      - Admin/Bands
    post:
      consumes:
      - application/json
      description: Create a new band with the provided details
      parameters:
      - description: Band details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/bands.CreateBandRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/bands.BandResponse'
        "400":
          description: Bad request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
      summary: Create a new band
      tags:
      - Admin/Bands
  /admin/v1/bands/{id}:
    delete:
      description: Delete an existing band
      parameters:
      - description: Band ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
      summary: Delete a band
      tags:
      - Admin/Bands
    get:
      description: Get details of a specific band
      parameters:
      - description: Band ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/bands.BandResponse'
        "400":
          description: Bad request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "404":
          description: Not found
          schema:
            type: string
      summary: Get a band by ID
      tags:
      - Admin/Bands
    put:
      consumes:
      - application/json
      description: Update an existing band with the provided details
      parameters:
      - description: Band ID
        in: path
        name: id
        required: true
        type: integer
      - description: Band details to update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/bands.UpdateBandRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/bands.BandResponse'
        "400":
          description: Bad request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "404":
          description: Not found
          schema:
            type: string
      summary: Update a band
      tags:
      - Admin/Bands
...
  /api/v1/account:
    put:
      consumes:
      - application/json
      description: Replace all fields of the current user's account with provided
        values
      parameters:
      - description: Full account details to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/accounts.UpdateAccountRequestPut'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/accounts.UpdateAccountResponse'
        "400":
          description: Invalid request body
          schema:
            type: string
        "401":
          description: Not authorized
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Update full account info
      tags:
      - Account
  /api/v1/account/photo:
    post:
      consumes:
      - multipart/form-data
      description: Upload a photo file for the current user
      parameters:
      - description: Photo file to upload
        in: formData
        name: photo
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad request
          schema:
            type: string
        "401":
          description: Not authorized
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Upload a photo
      tags:
      - Account
...
  /v1/account:
    get:
      description: Return information about current user
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/accounts.GetAccountResponse'
        "401":
          description: Not authorized
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Get account info
      tags:
      - Account
    patch:
      consumes:
      - application/json
      description: Update specific fields of the current user's account
      parameters:
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/accounts.UpdateAccountRequestPatch'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/accounts.UpdateAccountResponse'
        "400":
          description: Invalid request body
          schema:
            type: string
        "401":
          description: Not authorized
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Update account info
      tags:
      - Account
swagger: "2.0"